genlifter: fix code generation for inline records
The main problem with inline records is that the absolute path to the
labels are prefixed, e.g.

  type t = A of { a : string; b : int }

generates code which looks like:

class virtual ['res] lifter =
  object (this)
    method lift_Test_t : Test.t -> 'res=
      (function
       | Test.A { Test.a = a; Test.b = b } ->
           this#constr "Test.t"
             ("A",
               (this#record "Test.t.A"
                  [("a", (this#string a)); ("b", (this#int b))])) : Test.t ->
                                                                    'res)
  end

which is not valid OCaml code:
for inline records, the path should just be an identifier
(disambiguation is based on the name of the constructor, e.g. Test.A in
this example).

This patch removes the prefix at two levels:
- qualified paths in the patterns in the lifter code
- qualified paths in the parsetree produced by the lifter