build under OCaml 4.08, 4.09, 4.10 and 4.11 from one repo
This switches the repository to be buildable using multiple
versions of OCaml, in a manner similar to that used by Merlin
and other compiler-dependent tools.

As a result, we can release a single ppx_tools package for
multiple compiler versions. This in turn makes it much easier
to use ppx_tools in dune monorepos, which makes life easier for
benchmarking systems such as Sandmark used in multicore OCaml.